requirements.txt
PyPDF2==3.0.1
langchain==0.2.11
langchain-community==0.2.5
langchain-core==0.2.23
langchain-google-genai==1.0.6
langchain-text-splitters==0.2.1
streamlit==1.29.0
faiss-gpu==1.7.2
reportlab==4.2.2
python-dotenv==1.0.1